Move follow requirements to new census_yourschoolv4 model
