+++ dashboard/test/models/census/census_your_school2017v1_test.rb
@@ -6,19 +6,9 @@ class Census::CensusYourSchool2017v1Test < ActiveSupport::TestCase
     assert submission.valid?, submission.errors.full_messages
   end
 
-  test "census 2017 v1 submission as teacher without pledge" do
-    submission = build(:census_your_school2017v1, :as_teacher)
-    assert_not submission.valid?, submission.errors.full_messages
-  end
-
   test "census 2017 v1 submission as teacher with pledge" do
     submission = build(:census_your_school2017v1, :as_teacher, :with_pledge)
     assert submission.valid?, submission.errors.full_messages
-  end
-
-  test "census 2017 v1 submission missing followup" do
-    submission = build(:census_your_school2017v1, :requiring_followup)
-    assert_not submission.valid?, submission.errors.full_messages
   end
 
   test "census 2017 v1 submission with followup" do
